Refactored folder structure, created API service file
