+++ package.json
@@ -24,6 +24,7 @@
     "primeng": "^13.1.0",
     "rxjs": "~7.4.0",
     "tslib": "^2.3.0",
+    "twitter-api-v2": "^1.10.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
@@ -39,5 +40,15 @@
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "~1.7.0",
     "typescript": "~4.4.3"
+  },
+  "browser": {
+    "crypto": false,
+    "http": false,
+    "https": false,
+    "net": false,
+    "fs": false,
+    "path": false,
+    "stream": false,
+    "tls": false
   }
 }
