+++ src/app/app.module.ts
@@ -4,13 +4,14 @@ import { BrowserModule } from '@angular/platform-browser';
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
-import { UserPanelComponent } from './user-panel/user-panel.component';
-import { SearchbarComponent } from './searchbar/searchbar.component';
-import { GalleryComponent } from './gallery/gallery.component';
+//import { UserPanelComponent } from './user-panel/user-panel.component';
+import { UserPanelComponent } from './components/user-panel/user-panel.component';
+import { SearchbarComponent } from './components/searchbar/searchbar.component';
+import { GalleryComponent } from './components/gallery/gallery.component';
 import { InputTextModule } from 'primeng/inputtext';
 import { Button, ButtonModule } from 'primeng/button';
 import { RippleModule } from 'primeng/ripple';
-import { Page } from './page';
+import { Page } from './components/page';
 import { DropdownModule } from 'primeng/dropdown';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { CalendarModule } from 'primeng/calendar';
