+++ src/app/components/gallery/gallery.component.css
