+++ src/app/components/gallery/gallery.component.spec.ts
