+++ src/app/components/gallery/gallery.component.ts
