+++ src/app/components/searchbar/searchbar.component.css
