+++ src/app/services/searchbar.service.ts
@@ -1,0 +1,13 @@
+import { Injectable } from '@angular/core';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class SearchbarService {
+
+  constructor() { }
+
+  getUserData() {
+    console.log("API Request")
+  }
+}
